extensions

+++ OxApp/console/SettingsTunnelOnServer.php
@@ -47,45 +47,46 @@ class SettingsTunnelOnServer extends Command
             Tunnels::where(['id' => $tunnel->id])->update([
                 'status' => 3,
             ]);
-            $server = Servers::find(['ip' => $tunnel->serverIp])->rows[0];
-            print_r($server);
-            $connection = ssh2_connect($server->ip, 22);
-            var_dump(ssh2_auth_password($connection, 'root', $server->password));
-            
-            ssh2_exec($connection, 'pkill 3proxy');
-            ssh2_exec($connection, 'ulimit -n 600000');
-            ssh2_exec($connection, 'ulimit -u 600000');
-            ssh2_exec($connection, 'ip link delete he-ipv6');
-            ssh2_exec($connection, 'ifconfig he-ipv6 down');
-            ssh2_exec($connection, 'ip -6 route del default');
-            ssh2_exec($connection, 'modprobe ipv6');
-            ssh2_exec($connection,
-                'ip tunnel add he-ipv6 mode sit remote ' . $tunnel->remoteIp . ' local ' . $tunnel->serverIp . ' ttl 255');
-            ssh2_exec($connection, 'ip link set he-ipv6 up');
-            ssh2_exec($connection, 'ip addr add ' . $tunnel->v6route . ' dev he-ipv6');
-            ssh2_exec($connection, 'ip route add ::/0 dev he-ipv6');
-            
-            $name = '48sub';
-            $exc = './fastProxy.sh ' . $tunnel->$name;
-            ssh2_exec($connection, $exc);
-            
-            $stream = ssh2_exec($connection, 'ifconfig');
-            stream_set_blocking($stream, true);
-            $stream_out = ssh2_fetch_stream($stream, SSH2_STREAM_STDIO);
-            echo stream_get_contents($stream_out);
-            
-            
-            Tunnels::where(['id' => $tunnel->id])->update([
-                'status' => 4,
-            ]);
-            Proxy::delete(['proxy/like' => $tunnel->serverIp . ':%']);
-            for ($i = 30000; $i < 30200; $i++) {
-                $proxy[] = Proxy::add([
-                    'proxy' => $tunnel->serverIp . ":" . $i . ";",
-                    'rand' => rand(0, 1000)
+            if ($tunnel->v6route !== '[1500]') {
+                $server = Servers::find(['ip' => $tunnel->serverIp])->rows[0];
+                print_r($server);
+                $connection = ssh2_connect($server->ip, 22);
+                var_dump(ssh2_auth_password($connection, 'root', $server->password));
+    
+                ssh2_exec($connection, 'pkill 3proxy');
+                ssh2_exec($connection, 'ulimit -n 600000');
+                ssh2_exec($connection, 'ulimit -u 600000');
+                ssh2_exec($connection, 'ip link delete he-ipv6');
+                ssh2_exec($connection, 'ifconfig he-ipv6 down');
+                ssh2_exec($connection, 'ip -6 route del default');
+                ssh2_exec($connection, 'modprobe ipv6');
+                ssh2_exec($connection,
+                    'ip tunnel add he-ipv6 mode sit remote ' . $tunnel->remoteIp . ' local ' . $tunnel->serverIp . ' ttl 255');
+                ssh2_exec($connection, 'ip link set he-ipv6 up');
+                ssh2_exec($connection, 'ip addr add ' . $tunnel->v6route . ' dev he-ipv6');
+                ssh2_exec($connection, 'ip route add ::/0 dev he-ipv6');
+    
+                $name = '48sub';
+                $exc = './fastProxy.sh ' . $tunnel->$name;
+                ssh2_exec($connection, $exc);
+    
+                $stream = ssh2_exec($connection, 'ifconfig');
+                stream_set_blocking($stream, true);
+                $stream_out = ssh2_fetch_stream($stream, SSH2_STREAM_STDIO);
+                echo stream_get_contents($stream_out);
+    
+    
+                Tunnels::where(['id' => $tunnel->id])->update([
+                    'status' => 4,
                 ]);
+                Proxy::delete(['proxy/like' => $tunnel->serverIp . ':%']);
+                for ($i = 30000; $i < 30200; $i++) {
+                    $proxy[] = Proxy::add([
+                        'proxy' => $tunnel->serverIp . ":" . $i . ";",
+                        'rand' => rand(0, 1000)
+                    ]);
+                }
             }
-            
             
         }
